feat: 참여한 모임 중 첫 번째 모임의 로그를 띄워주는 기능 구현

+++ android/app/src/main/java/com/woowacourse/ody/data/remote/service/NotificationService.kt
@@ -7,7 +7,7 @@ import retrofit2.http.Path
 interface NotificationService {
     @GET(NOTIFICATION_LOG_PATH)
     suspend fun fetchNotificationLogs(
-        @Path("meetingId") meetingId: Int,
+        @Path("meetingId") meetingId: Long,
     ): NotificationLogsResponse
 
     companion object {
